+++ unreal/AegisWar/Source/AegisWar/Public/WarPlayerController.h
@@ -6,6 +6,7 @@
 
 class UWarEntryStatusWidget;
 class UWarInventoryWidget;
+class UWarQuestLogWidget;
 
 UCLASS()
 class AEGISWAR_API AWarPlayerController : public APlayerController
@@ -17,10 +18,12 @@ public:
     const FText& GetEntryFailure() const { return LastEntryFailure; }
     virtual void SetupInputComponent() override;
     void ToggleInventory();
+    void ToggleQuestLog();
     void InteractWithStation();
     void InteractWithWorld();
 private:
     UPROPERTY(Transient) TObjectPtr<UWarInventoryWidget> InventoryWidget;
+    UPROPERTY(Transient) TObjectPtr<UWarQuestLogWidget> QuestLogWidget;
     UPROPERTY(Transient) TObjectPtr<UWarEntryStatusWidget> EntryStatus;
     UPROPERTY(Transient) FText LastEntryFailure;
 };